feat: debugging simd2 segfaults
The current code that is being generated is correct (I think), but I am concerned
the code generated by the compiler is not correct. GDB is not showing
valin and valout variables on the stack for storing data back to the
array. We are getting the actual segfault on _mm_store_ps(&out[9], valout2);
And when you try and print valout2, gdb returns that it has been
"optimized out". So that's worrying. More investigation is required.